revert move _SecretManagementAttributes to _internal
