+++ azure-keyvault/azure/keyvault/_internal.py
@@ -58,44 +58,3 @@ class _BackupResult(Model):
     def __init__(self, **kwargs):
         super(_BackupResult, self).__init__(**kwargs)
         self.value = None
-
-
-class _SecretManagementAttributes(Model):
-    """The secret management attributes.
-
-    Variables are only populated by the server, and will be ignored when
-    sending a request.
-
-    :param bool enabled: Determines whether the object is enabled.
-    :param datetime not_before: Not before date in UTC.
-    :param datetime expires: Expiry date in UTC.
-    :ivar datetime created: Creation time in UTC.
-    :ivar datetime updated: Last updated time in UTC.
-    :ivar str recovery_level: Reflects the deletion recovery level currently in
-     effect for secrets in the current vault. If it contains 'Purgeable', the
-     secret can be permanently deleted by a privileged user; otherwise, only
-     the system can purge the secret, at the end of the retention interval.
-     Possible values include: 'Purgeable', 'Recoverable+Purgeable',
-     'Recoverable', 'Recoverable+ProtectedSubscription'
-    """
-
-    _validation = {"created": {"readonly": True}, "updated": {"readonly": True}, "recovery_level": {"readonly": True}}
-
-    _attribute_map = {
-        "enabled": {"key": "enabled", "type": "bool"},
-        "not_before": {"key": "nbf", "type": "unix-time"},
-        "expires": {"key": "exp", "type": "unix-time"},
-        "created": {"key": "created", "type": "unix-time"},
-        "updated": {"key": "updated", "type": "unix-time"},
-        "recovery_level": {"key": "recoveryLevel", "type": "str"},
-    }
-
-    def __init__(self, **kwargs):
-        # type: (Mapping[str, Any]) -> None
-        super(_SecretManagementAttributes, self).__init__(**kwargs)
-        self.enabled = kwargs.get("enabled", None)
-        self.not_before = kwargs.get("not_before", None)
-        self.expires = kwargs.get("expires", None)
-        self.created = None
-        self.updated = None
-        self.recovery_level = None
